+++ packages/components/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+
+- `Tree / TreeItem` disabled and selected states
+
+### Fixed
+
+- `TooltipContent` default width is back to `'auto'`
+- Erratic scrolling after dynamic list resize in all `Combobox`-based components
+
 ## [0.9.29]
 
 ### Fixed